refactor: Test에 필요한 Fake 객체들의 파일 분리
