+++ app/src/test/java/woowacourse/shopping/FakeUIObjects.kt
@@ -1,0 +1,17 @@
+package woowacourse.shopping
+
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModel
+import woowacourse.shopping.ui.util.DependencyInjector
+import woowacourse.shopping.ui.util.ReflectiveViewModelFactory
+
+class FakeViewModel(
+    val fakeRepository: FakeRepository,
+) : ViewModel()
+
+class FakeActivity : AppCompatActivity() {
+    val viewModel: FakeViewModel by viewModels {
+        ReflectiveViewModelFactory(DependencyInjector(fakeDependencyContainer))
+    }
+}
